nautobot-plugin-data-sync-servicenow/pyproject.toml
[tool.poetry]
name = "nautobot-data-sync-servicenow"
version = "0.1.0"
description = "Nautobot ServiceNow Data Synchronization"
authors = ["Network to Code, LLC <opensource@networktocode.com>"]

license = "Apache-2.0"

readme = "README.md"
homepage = "https://github.com/nautobot/nautobot-plugin-data-sync-servicenow"
repository = "https://github.com/nautobot/nautobot-plugin-data-sync-servicenow"
keywords = ["nautobot", "nautobot-plugin"]
include = [

    "LICENSE",

    "README.md",
]
packages = [
    { include = "nautobot_data_sync_servicenow" },
]

[tool.poetry.dependencies]
python = "^3.6"
Jinja2 = "<3"
pysnow = "^0.7.17"
usaddress = "^0.5.10"
PyYAML = "^5.4.1"
diffsync = "^1.3.0"

[tool.poetry.dev-dependencies]
invoke = "*"
black = "*"
django-debug-toolbar = "*"
yamllint = "*"
bandit = "*"
pylint = "*"
pylint-django = "*"
pydocstyle = "*"
flake8 = "*"
coverage = "*"
mkdocs = "*"
markdown-include = "*"

[tool.poetry.plugins."nautobot_data_sync.sync_worker_classes"]
"ServiceNow export" = "nautobot_data_sync_servicenow.worker:ServiceNowExportDataSyncWorker"

[tool.black]
line-length = 120
target-version = ['py37']
include = '\.pyi?$'
exclude = '''
(
  /(
      \.eggs         # exclude a few common directories in the
    | \.git          # root of the project
    | \.hg
    | \.mypy_cache
    | \.tox
    | \.venv
    | _build
    | buck-out
    | build
    | dist
  )/
  | settings.py     # This is where you define files that should not be stylized by black
                     # the root of the project
)
'''

[tool.pylint.master]
# Include the pylint_django plugin to avoid spurious warnings about Django patterns
load-plugins="pylint_django"

[tool.pylint.basic]
# No docstrings required for private methods (Pylint default), or for test_ functions, or for inner Meta classes.
no-docstring-rgx="^(_|test_|Meta$)"

[tool.pylint.messages_control]
# Line length is enforced by Black, so pylint doesn't need to check it.
# Pylint and Black disagree about how to format multi-line arrays; Black wins.
disable = """,
    line-too-long,
    bad-continuation,
    """

[tool.pylint.miscellaneous]
# Don't flag TODO as a failure, let us commit with things that still need to be done in the code
notes = """,
    FIXME,
    XXX,
    """

[build-system]
requires = ["poetry_core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
testpaths = [
    "tests"
]
addopts = "-vv --doctest-modules"
